Add unit test for template function
